% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plus.R
\name{compare_plus}
\alias{compare_plus}
\title{compare_plus}
\usage{
compare_plus(models, discount, model_prior = NULL, state_prior = NULL, obs,
  action, alphas = NULL, verbose = TRUE, mc.cores = 1L, ...)
}
\arguments{
\item{models}{a list of lists, each of which gives the transition matrix, observation matrix and reward matrix for the model considered}

\item{discount}{discount rate}

\item{model_prior}{Prior belief assigned to each model. uniform by default.}

\item{state_prior}{Prior belief that system is in state x_i before simulation starts}

\item{obs}{vector of observations from given data}

\item{alphas}{the alpha vectors for each model, as provided from \code{\link{sarsop_plus}}, which will otherwise be run each time if not provided.}

\item{verbose}{should function provide messages?}

\item{mc.cores}{number of parallel cores. Only relevant if alphas not provided. Be sure adequate memory is available first!}

\item{...}{additiional options to appl::sarsop, if alphas are not provided}

\item{vector}{of actions corresponding to each observation, given the data}
}
\value{
a list with elements
(1) \code{df} a data.frame with the time, state, observation, action, and value at each timestep,
(2) \code{model_posterior}, a data.frame of Tmax rows by n_models columns giving the evolution of the belief over models, and
(3) \code{state_posterior}, a data.frame of Tmax rows by n_states columns giving the evolution of the belief over states
}
\description{
compare_plus
}
\examples{
\dontrun{
source(system.file("examples/K_models.R", package="pomdpplus"))
alphas <- sarsop_plus(models, discount, precision = 1)
out <- compare_plus(models = models, discount = discount,
                obs = rnorm(21, 15, 0.1),
                action = rep(1, 21),
                alphas = alphas)
}

}

