% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_models.R
\name{init_models}
\alias{init_models}
\title{function for generating approimation of value function for each candidate model POMDP}
\usage{
init_models(T, O, R, GAMMA, Num_model = length(T), initial = NULL,
  precision = 0.001, timeout = 1000)
}
\arguments{
\item{T}{list of transition probabilities matrices for all candidate models; length = Num_model}

\item{O}{list of emission probabilities matrices for all candidate models; length = Num_model}

\item{R}{matrix of reward function}

\item{GAMMA}{discount factor}

\item{Num_model}{number of candidate models}

\item{initial}{initial belief over the states}

\item{precision}{precision to run the models}

\item{timeout}{if you want to stop the run based on a specific time}
}
\value{
av list of alpha vectors for all candidate models; length = Num_Model

aa list of actions corresponding to alpha vectors for all candidate models; length = Num_Model
}
\description{
function for generating approimation of value function for each candidate model POMDP
}

