% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_replicate.R
\name{plus_replicate}
\alias{plus_replicate}
\title{plus_replicate}
\usage{
plus_replicate(reps, expr, ...)
}
\arguments{
\item{reps}{number of replicates to simulate}

\item{expr}{expression to replicate; must return a}

\item{...}{additional arguments to \code{\link{mclapply}}, such as mc.cores}
}
\value{
Same structure as returned by the given expression (e.g. sim_plus),
 but with replicates appended as additional rows, with an indicator column
}
\description{
Replicate sim_plus calls, potentially in parallel. Can also replicate mdp_learning
and mdp_planning simulation calls
}
