% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_seq.R
\name{plus_seq}
\alias{plus_seq}
\title{main code for planning and learning in uncertain systems}
\usage{
<<<<<<< HEAD
plus_seq(input, Num_sim = 100, seq, n_sample = 5, initial, P = (array(1,
  dim = length(input[[1]]))/length(input[[1]])))
=======
plus_seq(input, t = 100, Num_sim = 100, seq, n_sample = 5, initial,
  P = (array(1, dim = length(input[[1]]))/length(input[[1]])))
>>>>>>> 5b4a900b986399d008454a55098fc4493f051c55
}
\arguments{
\item{input}{A list of transition (itself a list for all candidate models),
emission (itslef a list for all candidate models), reward matrix, and discount factor.}

<<<<<<< HEAD
=======
\item{t}{time horizon of the problem ; default = 100}

>>>>>>> 5b4a900b986399d008454a55098fc4493f051c55
\item{Num_sim}{Number of simulations replicates ; default = 100}

\item{seq}{list containing the sequnce of actions and observations from environment}

\item{n_sample}{number of sample to use for planning under model uncertainty; default = 5}

\item{initial}{initial belief state}

\item{P}{prior probability of the models ; default is flat prior}
}
\value{
df data farme including results of forward simulations

posterior data frame of Posterior distribution of each candidate model at each time

av list of alpha vectors for all candidate models; length = Num_Model

aa list of actions corresponding to alpha vectors for all candidate models; length = Num_Model
}
\description{
main code for planning and learning in uncertain systems
}

