% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{function for running forward simulations for planning and learning under model uncertainty}
\usage{
run_sim(T, O, R, GAMMA, av, aa, n, Num_sim, t, N = length(T), init, n_sample,
  P = (array(1, dim = c(1, length(T)))/length(T)))
}
\arguments{
\item{T}{list of transition probabilities matrices for all candidate models; length = Num_model}

\item{O}{list of emission probabilities matrices for all candidate models; length = Num_model}

\item{R}{matrix of reward function}

\item{GAMMA}{discount factor}

\item{av}{list of alpha vectors for all candidate models; length = Num_Model}

\item{aa}{list of actions corresponding to alpha vectors for all candidate models; length = Num_Model}

\item{n}{index of the true model}

\item{Num_sim}{number of simulation replicates; default = 100}

\item{t}{time horizon of the simulaitons; default = 100}

\item{N}{number of candidate models}

\item{init}{initial belief of the states}

\item{n_sample}{optional number of sample to use for planning under model uncertainty; If not specified, all model will be used}

\item{P}{prior probability of the models ; default is flat prior}
}
\value{
df data farme including results of forward simulations

posterior data frame of Posterior distribution of each candidate model at each time
}
\description{
function for running forward simulations for planning and learning under model uncertainty
}

