% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_seq.R
\name{run_sim_seq}
\alias{run_sim_seq}
\title{function for running forward simulations for planning and learning under model uncertainty}
\usage{
run_sim_seq(T, O, R, GAMMA, av, aa, Num_sim, seq, N = length(T), init,
  n_sample, P = (array(1, dim = c(1, length(T)))/length(T)))
}
\arguments{
\item{T}{list of transition probabilities matrices for all candidate models; length = Num_model}

\item{O}{list of emission probabilities matrices for all candidate models; length = Num_model}

\item{R}{matrix of reward function}

\item{GAMMA}{discount factor}

\item{av}{list of alpha vectors for all candidate models; length = Num_Model}

\item{aa}{list of actions corresponding to alpha vectors for all candidate models; length = Num_Model}

\item{Num_sim}{number of simulation replicates; default = 100}

\item{seq}{list containing the sequnce of actions and observations from environment}

\item{N}{number of candidate models}

\item{init}{initial belief of the states}

\item{n_sample}{number of samples used for planning; default = 5}

\item{P}{prior probability of the models ; default is flat prior}

\item{t}{time horizon of the simulaitons; default = 100}
}
\value{
df data farme including results of forward simulations

posterior data frame of Posterior distribution of each candidate model at each time
}
\description{
function for running forward simulations for planning and learning under model uncertainty
}

