% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_plus.R
\name{sim_plus}
\alias{sim_plus}
\title{sim_plus}
\usage{
sim_plus(models, discount, model_prior = NULL, state_prior = NULL, x0,
  a0 = 1, Tmax, true_model, alphas = NULL, model_names = NULL,
  state_names = NULL, ...)
}
\arguments{
\item{models}{a list of lists, each of which gives the transition matrix, observation matrix and reward matrix for the model considered}

\item{discount}{discount rate}

\item{model_prior}{Prior belief assigned to each model. uniform by default.}

\item{state_prior}{Prior belief that system is in state x_i before simulation starts}

\item{x0}{starting state of the system}

\item{a0}{initial action (used to make the first observation, only relevant if observation depends on action)}

\item{Tmax}{number of time steps to simulate}

\item{true_model}{a list of the transition matrix, observation matrix, and reward matrix used to simulate draws.}

\item{alphas}{the alpha vectors for each model, as provided from \code{\link{sarsop_plus}}, which will otherwise be run each time if not provided.}

\item{model_names}{vector of identifying names for each model. If none are provided, model posterior columns will be named V1, V2, etc.}

\item{state_names}{vector of identifying names for each state.  If none are provided, state posterior columns will be named V1, V2, etc.}

\item{...}{additional options to appl::sarsop, if alphas are not provided}
}
\value{
a list with elements
(1) \code{df} a data.frame with the time, state, observation, action, and value at each timestep,
(2) \code{model_posterior}, a data.frame of Tmax rows by n_models columns giving the evolution of the belief over models, and
(3) \code{state_posterior}, an array of Tmax rows by n_states by n_models columns giving the evolution of the belief over states
}
\description{
sim_plus
}
\examples{
\dontrun{
source(system.file("examples/K_models.R", package="pomdpplus"))
alphas <- sarsop_plus(models, discount, precision = 1)
out <- sim_plus(models = models, discount = discount,
                x0 = 5, a0 = 1, Tmax = 10,
                true_model = models[[2]], alphas = alphas)
}

}

