% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_seq.R
\name{plus_seq}
\alias{plus_seq}
\title{main code for planning and learning in uncertain systems}
\usage{
plus_seq(input, Num_sim = 100, seq, n_sample, initial, P = (array(1, dim =
  length(input[[1]]))/length(input[[1]])), precision = 0.001)
}
\arguments{
\item{input}{A list of transition (itself a list for all candidate models),
emission (itslef a list for all candidate models), reward matrix, and discount factor.}

\item{Num_sim}{Number of simulations replicates ; default = 100}

\item{seq}{list containing the sequnce of actions and observations from environment}

\item{n_sample}{optional number of sample to use for planning under model uncertainty; If not specified, all model will be used}

\item{initial}{initial belief state}

\item{P}{prior probability of the models ; default is flat prior}

\item{precision}{accuracy of value function for generating models}
}
\value{
df data farme including results of forward simulations

posterior data frame of Posterior distribution of each candidate model at each time

av list of alpha vectors for all candidate models; length = Num_Model

aa list of actions corresponding to alpha vectors for all candidate models; length = Num_Model
}
\description{
main code for planning and learning in uncertain systems
}

